chore: Simplify the app and cleanup
