+++ .claude-plugin/marketplace.json
@@ -6,7 +6,7 @@
   "plugins": [
     {
       "name": "repo-indexer",
-      "version": "0.0.3",
+      "version": "0.0.4",
       "source": "./",
       "description": "Indexes codebases for persistent Claude context with tiered memory architecture"
     }
