--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,32 +1,0 @@
----
-name: Bug report
-about: Something isn't working correctly
-labels: bug
----
-
-## Describe the bug
-
-<!-- A clear and concise description of what went wrong -->
-
-## Environment
-
-- **Repo type:** <!-- monorepo / microservices / single_app / library -->
-- **Phase where failure occurred:** <!-- 1 (git sync) / 2 (detect) / 3 (index) / 4 (generate) / 5 (validate) / 6 (memory) -->
-- **Python version:** <!-- e.g. 3.11.4 -->
-- **OS:** <!-- e.g. macOS 14, Ubuntu 22.04 -->
-
-## Script output
-
-<!-- Run the relevant script manually and paste the output -->
-
-```bash
-python3 skills/repo-indexer/scripts/<script>.py
-```
-
-## Expected behavior
-
-<!-- What did you expect to happen? -->
-
-## Additional context
-
-<!-- Anything else that might help diagnose the issue -->
