--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,25 +1,0 @@
----
-name: Feature request
-about: Suggest an improvement or new capability
-labels: enhancement
----
-
-## Use case
-
-<!-- What are you trying to accomplish? Who benefits? -->
-
-## Current behavior
-
-<!-- What happens today that makes this difficult or impossible? -->
-
-## Proposed solution
-
-<!-- What would you like to see? How should it work? -->
-
-## Alternatives considered
-
-<!-- Any other approaches you've thought about? -->
-
-## Additional context
-
-<!-- Mockups, examples from other tools, related issues, etc. -->
