+++ .github/workflows/ci.yml
@@ -53,16 +53,6 @@ jobs:
           flake8 skills/repo-indexer/scripts/ \
             --max-line-length=120
 
-  lint-shell:
-    name: Lint Shell
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-      - name: Install shellcheck
-        run: sudo apt-get install -y shellcheck
-      - name: Run shellcheck
-        run: shellcheck skills/repo-indexer/scripts/git-sync.sh
-
   version-sync:
     name: Version Sync
     runs-on: ubuntu-latest
