+++ ruff.toml
@@ -1,0 +1,8 @@
+line-length = 120
+
+[lint]
+select = ["E", "F", "W", "I", "N", "UP", "S"]
+ignore = ["S603"]
+
+[lint.per-file-ignores]
+"tests/*" = ["S603", "S101"]
