+++ skills/repo-indexer/scripts/detect-repo-type.py
@@ -154,36 +154,6 @@ def detect_repo_type(root: str = ".") -> dict:
         indicators["microservices"] += len(dockerfiles)
         evidence.append(f"{len(dockerfiles)} Dockerfiles found")
 
-    # Gradle multi-project: settings file is the authoritative signal
-    gradle_settings = next(
-        (f for f in ("settings.gradle", "settings.gradle.kts") if (path / f).exists()),
-        None,
-    )
-    if gradle_settings:
-        indicators["monorepo"] += _MONOREPO_CONFIG_SCORE
-        evidence.append(f"Found {gradle_settings} (Gradle multi-project)")
-        # Supporting build files only counted when settings file exists
-        if any((path / f).exists() for f in ("build.gradle", "gradlew")):
-            indicators["monorepo"] += 1
-            evidence.append("Found build.gradle/gradlew")
-
-    # Check for Bazel workspace (WORKSPACE file is the authoritative signal)
-    bazel_workspace = next(
-        (
-            f
-            for f in ("WORKSPACE", "WORKSPACE.bazel", "MODULE.bazel")
-            if (path / f).exists()
-        ),
-        None,
-    )
-    if bazel_workspace:
-        indicators["monorepo"] += _MONOREPO_CONFIG_SCORE
-        evidence.append(f"Found {bazel_workspace} (Bazel)")
-        # Supporting build files only counted when workspace file exists
-        if any((path / f).exists() for f in ("BUILD", ".bazelrc")):
-            indicators["monorepo"] += 1
-            evidence.append("Found BUILD/.bazelrc")
-
     # Check for library indicators
     lib_markers = [
         "setup.py",
