+++ skills/repo-indexer/scripts/estimate-tokens.py
@@ -4,48 +4,6 @@
 import sys
 from pathlib import Path
 
-# Bytes-per-token ratios by content mode.
-# "prose" and "default" use 4 bytes/token; "code" uses 3 (denser tokenization).
-_BYTES_PER_TOKEN = {"default": 4, "prose": 4, "code": 3}
-
-# File extensions typically containing source code (denser tokenization)
-_CODE_EXTENSIONS = {
-    ".py",
-    ".js",
-    ".ts",
-    ".jsx",
-    ".tsx",
-    ".go",
-    ".rs",
-    ".java",
-    ".c",
-    ".cpp",
-    ".h",
-    ".hpp",
-    ".cs",
-    ".rb",
-    ".php",
-    ".swift",
-    ".kt",
-    ".scala",
-    ".sh",
-    ".bash",
-    ".zsh",
-    ".fish",
-    ".sql",
-    ".graphql",
-    ".proto",
-    ".yaml",
-    ".yml",
-    ".toml",
-    ".json",
-    ".xml",
-    ".html",
-    ".css",
-    ".scss",
-    ".sass",
-}
-
 # Aggregate budget for all L2 memory files combined
 L2_TOTAL_BUDGET = 10_000
 
@@ -59,28 +17,13 @@ BUDGETS = {
 # Default budget applied to any memory file not listed above
 MEMORY_DEFAULT_BUDGET = 5000
 
-
-def _guess_content_mode(filepath: "Path") -> str:
-    """Return 'code' for source files, 'prose' otherwise."""
-    if filepath.suffix.lower() in _CODE_EXTENSIONS:
-        return "code"
-    return "prose"
-
-
-def estimate_tokens(text: str, mode: str = "default") -> int:
-    """Convert UTF-8 byte length to an approximate token count.
-
-    Args:
-        text: The text to estimate.
-        mode: 'default' or 'prose' (4 bytes/token), 'code' (3 bytes/token).
-              Unknown modes fall back to 'default'.
-    """
-    bpt = _BYTES_PER_TOKEN.get(mode, _BYTES_PER_TOKEN["default"])
-    return len(text.encode("utf-8")) // bpt
-
-
 # Skip files larger than this to avoid reading multi-GB files into memory
 _MAX_FILE_BYTES = 1_000_000  # 1 MB
+
+
+def estimate_tokens(text: str) -> int:
+    """Convert UTF-8 byte length to an approximate token count (4 bytes/token)."""
+    return len(text.encode("utf-8")) // 4
 
 
 def check_file(filepath: Path) -> dict:
@@ -119,8 +62,7 @@ def check_file(filepath: Path) -> dict:
             "over": True,
             "pct": None,
         }
-    mode = _guess_content_mode(filepath)
-    tokens = estimate_tokens(content, mode=mode)
+    tokens = estimate_tokens(content)
     return {
         "exists": True,
         "tokens": tokens,
